requirements.txt
anyio==4.9.0
certifi==2025.1.31
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.4
opencv-python==4.11.0.86
pdf2image==1.17.0
pillow==11.1.0
PyPDF2==3.0.1
python-telegram-bot==22.0
sniffio==1.3.1
typing_extensions==4.13.0
